Replace double with BigDecimal and centralize hardcoded strings
- Remove unused jsoniter dependency (Jackson only requirement)
- Replace all double monetary fields with BigDecimal for precision
- Create ApiConstants for Fabrick API endpoint URLs
- Create EventBusConstants for event bus addresses
- Create StatusConstants for status codes and error messages
- Update all verticles to use new constants classes
- Update unit tests for BigDecimal comparisons

+++ src/main/java/it/demo/fabrick/dto/BalanceDto.java
@@ -1,5 +1,6 @@
 package it.demo.fabrick.dto;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 
 import lombok.Data;
@@ -14,8 +15,8 @@ public class BalanceDto {
 	@Data
 	public class Payload{
 	    public String date;
-	    public double balance;
-	    public double availableBalance;
+	    public BigDecimal balance;
+	    public BigDecimal availableBalance;
 	    public String currency;
 	}
